Remove non-model creation from models.py to respective files.
Registering -> Registering of models (reversion, watson) are done in appconfig.ready(), when Django says the application is ready.
Signals -> signals.py (hooks, e.g. post_save, post-delete)
Tasks -> tasks.py (tasks which modifies a model instance)
