+++ apps/offline/models.py
@@ -1,16 +1,11 @@
 # -*- coding: utf-8 -*-
 
-import logging
 from os import path
-from subprocess import check_call, CalledProcessError
 from django.db import models
-from django.db.models.signals import post_save
 from django.utils.translation import ugettext_lazy as _
 from filebrowser.fields import FileBrowseField
 from onlineweb4.settings.local import MEDIA_ROOT
 from chunks.models import Chunk
-
-import reversion
 
 THUMBNAIL_HEIGHT = 200  # Ønsket høyde på thumbnail
 IMAGE_FOLDER = "images/offline"
@@ -21,9 +16,6 @@ class ProxyChunk(Chunk):
         proxy = True
         verbose_name = 'Informasjonstekst'
         verbose_name_plural = 'Informasjonstekster'
-
-
-reversion.register(Chunk)
 
 
 class Issue(models.Model):
@@ -75,36 +67,3 @@ class Issue(models.Model):
         permissions = (
             ('view_issue', 'View Issue'),
         )
-
-
-reversion.register(Issue)
-
-
-def create_thumbnail(sender, instance=None, **kwargs):
-    logger = logging.getLogger(__name__)
-    logger.debug('Checking for thumbnail...')
-    if instance is None:
-        return
-    t = Issue.objects.get(id=instance.id)
-
-    if t.thumbnail_exists is False:
-        logger.debug('Thumbnail not found - creating...')
-
-        # Fixes an annoying Exception in logs, not really needed
-        # http://stackoverflow.com/questions/13193278/ {
-        import threading
-        threading._DummyThread._Thread__stop = lambda x: 42
-        # }
-
-        try:
-            check_call(["convert", "-resize", "x" + str(THUMBNAIL_HEIGHT), t.url + "[0]", t.thumbnail])
-        except (OSError, CalledProcessError) as e:
-            logger.debug("ERROR: {0}".format(e))
-
-        logger.debug('Thumbnail created, and is located at: %s' % t.thumbnail)
-
-    else:
-        logger.debug('Thumbnail already exists, and is located at: %s' % t.thumbnail)
-
-
-post_save.connect(create_thumbnail, sender=Issue)
