+++ apps/profiles/appconfig.py
@@ -1,5 +1,7 @@
 from django.apps import AppConfig
 
+import reversion
+
 
 class ProfilesConfig(AppConfig):
     name = 'apps.profiles'
@@ -7,3 +9,7 @@ class ProfilesConfig(AppConfig):
 
     def ready(self):
         super(ProfilesConfig, self).ready()
+
+        from apps.profiles.models import Privacy
+
+        reversion.register(Privacy)
